refactor(withdraw-collateral): add setAmount to useWithdrawCollateral hook
The `setAmount` function is now passed to the `useWithdrawCollateral` hook to reset the amount input field after a successful or failed withdrawal. This improves user experience by clearing the input field post-transaction. Additionally, the `refetchHelthFactor` function is refetched in the `useBorrowOverview` hook when `collateralData` changes to ensure the health factor is up-to-date.